# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CamMonSys API
  version: 1.0.0
  description: Role-based camera monitoring REST and Server-Sent Events API.
servers:
  - url: /
    description: Same-origin application server
security:
  - cookieAuth: []
...
paths:
...
  /api/devices:
    get:
      tags: [Devices]
      summary: List active devices
      description: Admin or Monitor only.
      responses:
        '200':
          description: Devices ordered newest first.
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Devices]
      summary: Register a device
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDeviceRequest' }
      responses:
        '201':
          description: Device registered.
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: '#/components/schemas/CreatedDevice' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Selected device group does not exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    get:
      tags: [Devices]
      summary: Get a device
      description: Admin/Monitor, or a Responder with an unresolved assignment for this device.
      responses:
        '200':
          description: Device details.
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: '#/components/schemas/DeviceSummary' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Devices]
      summary: Soft-delete a device
      description: Admin only. Rejected while unresolved events exist.
      responses:
        '204': { description: Device soft-deleted. }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Device has unresolved events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/group:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    patch:
      tags: [Devices, Device groups]
      summary: Assign or remove a device group
      description: Admin only. A null groupId makes the device ungrouped.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [groupId]
              properties:
                groupId:
                  oneOf:
                    - { type: string, format: uuid }
                    - { type: 'null' }
      responses:
        '200':
          description: Assignment updated.
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: '#/components/schemas/DeviceGroupAssignment' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Device or group does not exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/start:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Devices]
      summary: Reserve a camera publication
      description: Admin only. Returns a publish-only LiveKit connection.
      responses:
        '200':
          description: Publication reserved.
          headers:
            Cache-Control: { schema: { type: string }, description: Always no-store. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CameraStart' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Device is unavailable or already reserved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/heartbeat:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Devices]
      summary: Renew a camera publication lease
      description: Admin only. The current authenticated session must own the publication.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishingSessionRequest' }
      responses:
        '200':
          description: Lease and Last Seen updated.
          content:
            application/json:
              schema:
                type: object
                required: [lastSeenAt, leaseExpiresAt]
                properties:
                  lastSeenAt: { type: string, format: date-time }
                  leaseExpiresAt: { type: string, format: date-time }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Publication is no longer active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/stop:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Devices]
      summary: Stop a camera publication
      description: Admin only. Releases the lease and queues room cleanup.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishingSessionRequest' }
      responses:
        '202':
          description: Publication released.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Publication is no longer current.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/view-token:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Devices]
      summary: Create a camera viewer connection
      description: Admin/Monitor, or a Responder with an unresolved device assignment. Returns a subscribe-only LiveKit connection.
      responses:
        '200':
          description: Viewer connection.
          headers:
            Cache-Control: { schema: { type: string }, description: Always no-store. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CameraConnection' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Camera is offline or access/publication changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/devices/{deviceId}/events:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    post:
      tags: [Events]
      summary: Create a simulated camera event
      description: Admin only. Requires ownership of the active publication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [publishingSessionId, type]
              properties:
                publishingSessionId: { type: string, format: uuid }
                type: { $ref: '#/components/schemas/EventType' }
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [publishingSessionId, type]
              properties:
                publishingSessionId: { type: string, format: uuid }
                type: { $ref: '#/components/schemas/EventType' }
                snapshot:
                  type: string
                  format: binary
                  description: Optional JPEG snapshot, at most 2 MiB.
      responses:
        '201':
          description: Event persisted.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event: { $ref: '#/components/schemas/CreatedEvent' }
        '400': { $ref: '#/components/responses/InvalidInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Snapshot exceeds the 2 MiB limit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: Publication is inactive.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cammon_token
      description: HttpOnly signed JWT backed by a revocable PostgreSQL session.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    InvalidInput:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Forbidden:
      description: The user lacks permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    EventType:
      type: string
      enum: [TEST_ALERT, MOTION, BED_EXIT]
    EventStatus:
      type: string
      enum: [OPEN, ACKNOWLEDGED, RESOLVED]
...
    CreatedDevice:
      type: object
      additionalProperties: false
      required: [id, name, location, created_by_id, group_id, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        created_by_id: { type: string, format: uuid }
        group_id: { $ref: '#/components/schemas/NullableUuid' }
        created_at: { type: string, format: date-time }
    DeviceSummary:
      type: object
      additionalProperties: false
      required: [id, name, location, group_id, last_seen_at, stream_version, created_at, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        group_id: { $ref: '#/components/schemas/NullableUuid' }
        last_seen_at: { $ref: '#/components/schemas/NullableDateTime' }
        stream_version:
          description: Current publishing session used to reconnect viewers when rooms change.
          $ref: '#/components/schemas/NullableUuid'
        created_at: { type: string, format: date-time }
        status: { type: string, enum: [ONLINE, OFFLINE] }
    CreateDeviceRequest:
      type: object
      additionalProperties: false
      required: [name, location]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        location: { type: string, minLength: 1, maxLength: 200 }
        groupId: { $ref: '#/components/schemas/NullableUuid' }
    DeviceGroupAssignment:
      type: object
      additionalProperties: false
      required: [id, group_id]
      properties:
        id: { type: string, format: uuid }
        group_id: { $ref: '#/components/schemas/NullableUuid' }
...
    PublishingSessionRequest:
      type: object
      additionalProperties: false
      required: [publishingSessionId]
      properties:
        publishingSessionId: { type: string, format: uuid }
    CameraConnection:
      type: object
      additionalProperties: false
      required: [serverUrl, token]
      properties:
        serverUrl: { type: string, format: uri, pattern: '^wss://' }
        token: { type: string, description: Short-lived LiveKit JWT. }
    CameraStart:
      allOf:
        - { $ref: '#/components/schemas/CameraConnection' }
        - type: object
          required: [publishingSessionId, heartbeatIntervalSeconds]
          properties:
            publishingSessionId: { type: string, format: uuid }
            heartbeatIntervalSeconds: { type: integer, minimum: 1 }
    CreatedEvent:
      type: object
      additionalProperties: false
      required: [id, device_id, type, status, created_at, assigned_to_id, snapshot_available, snapshot_captured_at]
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/EventType' }
        status: { $ref: '#/components/schemas/EventStatus' }
        created_at: { type: string, format: date-time }
        assigned_to_id: { $ref: '#/components/schemas/NullableUuid' }
        snapshot_available: { type: boolean }
        snapshot_captured_at: { $ref: '#/components/schemas/NullableDateTime' }
...
    NullableUuid:
      oneOf:
        - { type: string, format: uuid }
        - { type: 'null' }
...
    NullableDateTime:
      oneOf:
        - { type: string, format: date-time }
        - { type: 'null' }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...